Refractered code for dahsboard UI to make it mroe organized
