+++ SacStateApp/src/components/ServicesList.js
@@ -1,0 +1,34 @@
+import React from 'react';
+import { View, Text, TouchableOpacity, Linking } from 'react-native';
+import styles from '../styles/ServicesStyles';
+
+const ServicesList = ({ services }) => {
+  const handlePress = (link) => {
+    if (link) {
+      Linking.openURL(link).catch((err) =>
+        console.error('Failed to open link:', err)
+      );
+    }
+  };
+
+  return (
+    <View style={styles.serviceContainer}>
+      <Text style={styles.servicesHeader}>Services Available</Text>
+      {services.length === 0 ? (
+        <Text style={styles.details}>No services available</Text>
+      ) : (
+        services.map((service, index) => (
+          <TouchableOpacity
+            key={index}
+            style={styles.serviceBox}
+            onPress={() => handlePress(service.service_link)}
+          >
+            <Text style={styles.serviceTitle}>{service.serv_name}</Text>
+          </TouchableOpacity>
+        ))
+      )}
+    </View>
+  );
+};
+
+export default ServicesList;
