+++ SacStateApp/src/styles/CalendarStyles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  calendarContainer: {
+    marginVertical: 20,
+    borderRadius: 10,
+    overflow: 'hidden',
+    elevation: 3,
+    marginBottom: 0,
+  },
+  calendar: {
+    borderWidth: 1,
+    borderColor: '#ddd',
+    borderRadius: 10,
+    backgroundColor: '#fff',
+  },
+});
+
+export default styles;
